chore: move PlushieBlock to bade/block
